+++ scripts/make_all_docs.sh
@@ -2,12 +2,14 @@
 
 FORCED=""
 IMGGEN=""
+TESTONLY=""
 FILES=""
 DISPMD=""
 for opt in "$@" ; do
   case $opt in
     -f ) FORCED=$opt ;;
     -i ) IMGGEN=$opt ;;
+    -t ) TESTONLY=$opt ;;
     -d ) DISPMD=$opt ;;
     -* ) echo "Unknown option $opt"; exit -1 ;;
     * ) FILES="$FILES $opt" ;;
@@ -32,11 +34,11 @@ rm -f tmpscad*.scad
 for lib in $PREVIEW_LIBS; do
     lib="$(basename $lib .scad)"
     mkdir -p images/$lib
-    if [ "$IMGGEN" != "" ]; then
+    if [ "$IMGGEN" != ""  -a "$TESTONLY" != ""]; then
         rm -f images/$lib/*.png images/$lib/*.gif
     fi
     echo "$lib.scad"
-    ../scripts/docs_gen.py ../$lib.scad -o $lib.scad.md -c $IMGGEN $FORCED -I images/$lib/ || exit 1
+    ../scripts/docs_gen.py ../$lib.scad -o $lib.scad.md -c $IMGGEN $FORCED $TESTONLY -I images/$lib/ || exit 1
     if [ "$DISPMD" != "" ]; then
         open -a Typora $lib.scad.md
     fi
